# api-gateway/api/docs/swagger.yaml
definitions:
...
  booking.ProviderServiceGetAllRes:
    properties:
      provider_services:
        items:
          $ref: '#/definitions/booking.ProviderServiceRes'
        type: array
    type: object
  booking.ProviderServiceGetByIdRes:
    properties:
      provider_service:
        $ref: '#/definitions/booking.ProviderServiceRes'
    type: object
  booking.ProviderServiceRes:
    properties:
      id:
        type: string
      provider_id:
        type: string
      service_id:
        type: string
      user_id:
        type: string
    type: object
  booking.ReviewCreateReq:
    properties:
      booking_id:
        type: string
      comment:
        type: string
      provider_id:
        type: string
      rating:
        type: integer
      user_id:
        type: string
    type: object
  booking.ReviewGetAllRes:
    properties:
      reviews:
        items:
          $ref: '#/definitions/booking.ReviewRes'
        type: array
    type: object
  booking.ReviewGetByIdRes:
    properties:
      review:
        $ref: '#/definitions/booking.ReviewRes'
    type: object
  booking.ReviewRes:
    properties:
      booking_id:
        type: string
      comment:
        type: string
      id:
        type: string
      provider_id:
        type: string
      rating:
        type: integer
      user_id:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /provider/review:
    post:
      consumes:
      - application/json
      description: Create a new review
      parameters:
      - description: Review data
        in: body
        name: review
        required: true
        schema:
          $ref: '#/definitions/booking.ReviewCreateReq'
      produces:
      - application/json
      responses:
        "200":
          description: Review created
          schema:
            type: string
        "400":
          description: Invalid request payload
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Create review
      tags:
      - review
  /provider/review/{id}:
    get:
      consumes:
      - application/json
      description: Get a review
      parameters:
      - description: Review ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/booking.ReviewGetByIdRes'
        "400":
          description: Invalid request payload
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get review
      tags:
      - review
    put:
      consumes:
      - application/json
      description: Update an existing review
      parameters:
      - description: Id
        in: query
        name: id
        type: string
      - description: Comment
        in: query
        name: comment
        type: string
      - description: Rating
        in: query
        name: rating
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Review updated
          schema:
            type: string
        "400":
          description: Invalid request payload
          schema:
            type: string
        "404":
          description: Review not found
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Update review
      tags:
      - review
  /provider/review/all:
    get:
      consumes:
      - application/json
      description: Get all review
      parameters:
      - description: Page
        in: query
        name: page
        type: integer
      - description: ProviderId
        in: query
        name: provider_id
        type: string
      - description: UserId
        in: query
        name: user_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/booking.ReviewGetAllRes'
        "400":
          description: Invalid parameters
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get all review
      tags:
      - review
  /provider/service:
    post:
      consumes:
      - application/json
      description: Create a new provider
      parameters:
      - description: ServiceId
        in: query
        name: service_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Provider created
          schema:
            type: string
        "400":
          description: Invalid request payload
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Create provider
      tags:
      - provider
  /provider/service/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a provider by ID
      parameters:
      - description: Provider ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Provider deleted
          schema:
            type: string
        "400":
          description: Invalid provider ID
          schema:
            type: string
        "404":
          description: Provider not found
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Delete provider
      tags:
      - provider
    get:
      consumes:
      - application/json
      description: Get a provider
      parameters:
      - description: Provider ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/booking.ProviderServiceGetByIdRes'
        "400":
          description: Invalid request payload
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get provider
      tags:
      - provider
  /provider/service/all:
    get:
      consumes:
      - application/json
      description: Get all provider
      parameters:
      - description: ProviderId
        in: query
        name: provider_id
        type: string
      - description: Page
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/booking.ProviderServiceGetAllRes'
        "400":
          description: Invalid parameters
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get all provider
      tags:
      - provider
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"